requirements.txt
aniso8601==9.0.1
transformers==4.33.0
timm
Pillow==10.0.0
requests==2.31.0
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
future==0.18.3
geographiclib==2.0
geopy==2.4.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pyap==0.3.1
PyPDF2==3.0.1
pytz==2023.3.post1
six==1.16.0
Werkzeug==2.3.7
numpy==1.23.5
pandas==1.5.3
opencv-python==4.8.0.76
pdf2image==1.16.3
PyMuPDF==1.23.5
PyMuPDFb==1.23.5
pypdfium2==4.21.0
ultralyticsplus